A hard performance improving block has been added to physics and particles. These systems will get better performance or else.
Former-commit-id: 5a1c0269efc2d32329be5177888ceccc3b68bd81

+++ Sample/source/TimeStepGuard.cpp
@@ -14,5 +14,5 @@ TimeStepGuard::TimeStepGuard(TimeStepTracker * tracker) {
 }
 TimeStepGuard::~TimeStepGuard() {
 	//Add a recommendation now
-	tracker->addRecommendation((float)(OS::Now()-startTime));
+	tracker->addRecommendation(OS::Now()-startTime);
 }
